📦  FEAT: [bulk/main] check files count + move types
