+++ csv-bulk-signature/src/types.ts
@@ -1,0 +1,21 @@
+import { LOGOS } from "./config";
+
+export interface Contact {
+  "Brand*": keyof typeof LOGOS;
+  "Full Name*": string;
+  Credentials: string;
+  "Title*": string;
+  "Office Phone*": string;
+  "Mobile Phone": string;
+  "Calendly Link": string;
+  skip?: boolean;
+}
+export interface TemplateData {
+  logoUrl: string;
+  fullName: string;
+  credentials: string;
+  title: string;
+  officePhone: string;
+  mobilePhone: string;
+  calendly: string;
+}
